Extracted shared table view offset coordination logic
